requirements.txt
setuptools
numpy
scikit-learn
matplotlib
torch
setuptools
nltk
pandas
scipy
tqdm
pyyaml
jupyterlab
pytorch-lightning
tensorboard
pytorch-nlp
